requirements.txt
dask[complete]
jupyter-server-proxy
notebook
pandas
fabric
spacy
Pillow
sentence_transformers